requirements.txt
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
Faker==28.1.0
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==1.7.1
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
toposort==1.10
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
